A Tragedy
Previously on arch I used sddm with the austronaut theme, but that for
the life of me I cant get it to work on nix. therefore I switched to
greetd with regreet as my greeter. I dont think I've said the word greet
this many times in one sentence, anyways I finnaly managed to figure out
how to theme and everything is fine now, it took a week but it worked.