A Time Of Peace.
After my numerous struggles with hyprland plugins and display managers I
began my journey on customizing neovim and zsh. I expected this battle
to be a hard  fought one but to my surprise an ally came to my aid, nvf
was a joy to fight alongside and our victory was swift, after witnessing
our display of power zsh gave up with ease. after a brief respite I
shall move on to conquoring waybar.